Unit Tests were added.

+++ loanApp-backend/egedemirbas/src/main/java/com/loanapp/egedemirbas/Credit/Entity/Credit.java
@@ -24,7 +24,7 @@ public class Credit implements Serializable {
     @GeneratedValue
     private Long id;
     private Date insDate;
-    private double creditAmount;
+    private float creditAmount;
     private Long userIdentityNumber;
     private Date userDateOfBirth;
     private EnumCreditResult creditResult;
